convert buffer helper to be promise based
