+++ lib/helpers/index.js
@@ -3,7 +3,7 @@
 module.exports = {
     defaultConfiguration   : require('./defaultConfiguration'),
     createConnectionString : require('./createConnectionString'),
-    convertToBuffer        : require('./convertToBuffer'),
+    convertToBufferAsync   : require('./convertToBufferAsync'),
     createTransactionId    : require('./createTransactionId'),
     exponentialBackoff     : require('./exponentialBackoff'),
     getPackageData         : require('./getPackageData'),
